progress on comments
