+++ edu.cuny.hunter.hybridize.tests/resources/HybridizeFunction/testPositionalParameters/in/A.py
@@ -1,8 +1,10 @@
 import tensorflow as tf
 
-@tf.function(None)
+
+@tf.function
 def test():
     pass
+
 
 if __name__ == '__main__':
     test()
