+++ edu.cuny.hunter.hybridize.tests/resources/HybridizeFunction/testPositionalParameters12/in/A.py
@@ -1,9 +1,10 @@
 import tensorflow as tf
 
-@tf.function(None, (tf.TensorSpec(shape=[None], dtype=tf.float32),), False, True, "google.matmul_low_rank_matrix")
-def test(x):
-    return x
+
+@tf.function(None, None, True, None, False, None, None, None, None, None)
+def test():
+    pass
+
 
 if __name__ == '__main__':
-  number = tf.constant([1.0, 1.0])
-  test(number)
+    test()
