+++ edu.cuny.hunter.hybridize.tests/resources/HybridizeFunction/testPositionalParameters13/in/A.py
@@ -1,0 +1,11 @@
+import tensorflow as tf
+
+
+@tf.function(None, (tf.TensorSpec(shape=[None], dtype=tf.float32),), False, True, "google.matmul_low_rank_matrix")
+def test(x):
+    return x
+
+
+if __name__ == '__main__':
+  number = tf.constant([1.0, 1.0])
+  test(number)
