+++ edu.cuny.hunter.hybridize.tests/resources/HybridizeFunction/testPositionalParameters2/in/A.py
@@ -1,9 +1,10 @@
 import tensorflow as tf
 
-@tf.function(None, (tf.TensorSpec(shape=[None], dtype=tf.float32),))
-def test(x):
-    return x
+
+@tf.function(None)
+def test():
+    pass
+
 
 if __name__ == '__main__':
-    number = tf.constant([1.0, 1.0])
-    print(test(number))
+    test()
