+++ edu.cuny.hunter.hybridize.tests/resources/HybridizeFunction/testPositionalParameters4/in/A.py
@@ -1,8 +1,10 @@
 import tensorflow as tf
 
-@tf.function(None, (tf.TensorSpec(shape=[None], dtype=tf.float32),), False, True)
+
+@tf.function(None, (tf.TensorSpec(shape=[None], dtype=tf.float32),), False)
 def test(x):
     return x
+
 
 if __name__ == '__main__':
     number = tf.constant([1.0, 1.0])
